Show errors in synthesis tab

+++ src/main.rs
@@ -32,8 +32,8 @@ pub struct Language {
     // synthesis tab
     graphemes: Vec<Grapheme>,
     new_grapheme: String,
-    max_syllables: (u8, u8),          // (function words, content words)
-    syllable_wgts: (Vec<u8>, Vec<u8>) // (function words, content words)
+    max_syllables: (u8, u8),            // (function words, content words)
+    syllable_wgts: (Vec<u16>, Vec<u16>) // (function words, content words)
 }
 
 impl Language {
